Restructured mesh sampler into class.
This was done so the mesh sampler and the utility functions/class for getting the closest points in each point cloud can both live in mesh sample while providing a clear distinction between them.
